delete movies function added

+++ src/main/java/com/outofbounds/empire/Movies/Models/Movie.java
@@ -34,7 +34,7 @@ public class Movie {
     @Column(name = "price")
     private int price;
 
-    public Movie(String title, String description, int ageLimit, String category, float price, String imageUrl) {
+    public Movie(String title, String description, int ageLimit, String category, int price, String imageUrl) {
         this.title = title;
         this.description = description;
         this.ageLimit = ageLimit;
